e2e-spring-2022: исправить board.page
